[WRFE-70](refactor): funnel window ReferenceError 수정
